Implement manifest get endpoint in controller
To ease the work operator has to do, this commit adds a new endpoint to
the controller that provides the manifest json file included in a
specific artefact. This way, the operator does not have to download an
artefact it is about to remove from a server, it can simply fetch the
files to be removed from the manifest of the artefact to remove.