Initial work towards first level un/installs
To ease setups and mutations down the line, marauder should be capable
of properly installing and uninstalling full artefact lines from a
server. Right now, only upgrading is possible (as it is the most common
usecase) but for the initial setup and further configuration down the
line, proper addition/removal logic for artefacts in general (not their
specific versions) is needed.