Allow marauder client to read config from stdin
To allow dynamic environments to easily invoke marauders client, this
commit adds the ability to pass the configuration yaml file to the
standard input of the marauder client.

For this, `-` (a single dash) is passed instead of a path to the
configuration flag of the client, instructing it to read the entire
standard input and interpreting it as the configuration yamls content.