Allow file matches to restrict matches
File matches defined in a marauder file match files by glob, allowing to
potentially match unwanted files if the build environment is not cleaned
properly.

To prevent unwanted artefacts from being created, this commit creates a
new json configuration in the file definition of a marauder manifest,
that allows definitions to limit the matches for the glob to a specific
number, usually to a single match for, e.g., plugin jars.