Enable updating without restarting
Some artefacts may safely be deployed to a server without said server
being restarted during the deployment.

To enable these seamless upgrades, this commit allows artefacts to
define themselves as not requiring a restart.
Marauder can then be instructed to execute a new lifecycle action on a
server, updateWithoutRestart, which only updates those artefacts that
can be updated (as per their manifest) without a server restart.