Move hashes into file reference
Moves the hashes of the files included in the artefact into the file
reference section of the manifest.

This, compared to the previous storage in a single giant manifest wide
map, allows later fetching of the file reference definition a file was
matched under.

This way, defined restrictions or future data specific to a file
definition in the manifest can be applied/read to the files matched
under it.