File equality rework (#4)
While the initial releases of marauder did not expect deployed files to
ever change, realistically that is not possible as some plugins and
their configuration files might run migrations and change the
configuration files, potentially even just reformatting files deployed
via marauder.

To allow marauder to still nicely deploy these files, marauder now
enables deployments to define an
`equalityProvider` for its deployed files.
Currently available are:

- `hash`
    Compares files based on their sha256sum hashes.
- `noop`
Compares files to be always equal. May be used if file content may
change.
- `json`
Compares files to be equal if their parsed json tree is equal. Comments
and format in the file are ignored.
- `yaml`
Compares files to be equal if their parsed yaml tree is equal. Comments
and format in the file are ignored.